+++ src/collective/documentgenerator/browser/generation_view.py
@@ -218,8 +218,6 @@ class DocumentGenerationView(BrowserView):
             forceOoCall=True,
             optimalColumnWidths=optimalColumnWidths,
             stylesMapping=stylesMapping,
-            managePageStyles=1,  # 'rename'
-            resolveFields=False,  # 'PageCount'
             **kwargs
         )